Extract circular queue fields/operations to a module
The Task had built-in circular queue.
I extracted the fields and operations to a type-generic, fixed-size RingQue.
All RingQue methods are templates so they will be built back into Task;
in other words, there should be no penalty for the calls to RingQue.

A little negative side effect is that qBuf.pop(), as a atomic template,
had to move its entirety inside the critical section.  Previously,
some was in, some was out.