Improvements to ringque
Added a clear() method.

Change queue index types from Natural to Positive/range.
This ensures a non-zero queue size.

Removed ref type since it is not needed at this time

Made the `=copy` proc an error to prevent implicit copying
since a RingQue is meant to be owned by a task.